`ModulePath` returns!
Replacing `ModulePath` with an array created an inconsistency in the API. For properties that could be either a Node/Token or an array, forced consumers to use `Array.isArray` instead of simply checking the `kind` property (as is done everywhere else in the API).

Adding a `kind` property to arrays returned from `parseList` was considered but was rejected.
* Adding a property to a built-in type can impact future compatibility.
* Having a dedicated type for module paths makes it easier to add methods specific to it.
* Default JSON serialization skips added properties. This isn't necessary for the language service to function but it is surprising behavior.